core: Move files to submodules
