+++ src/core/bond_breakage/CMakeLists.txt
@@ -1,0 +1,2 @@
+target_sources(EspressoCore
+               PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/bond_breakage.cpp)
