+++ src/script_interface/lees_edwards/OscillatoryShear.hpp
@@ -19,7 +19,7 @@
 #ifndef SCRIPT_INTERFACE_LEES_EDWARDS_OSCILLATORY_SHEAR_HPP
 #define SCRIPT_INTERFACE_LEES_EDWARDS_OSCILLATORY_SHEAR_HPP
 
-#include "core/lees_edwards.hpp"
+#include "core/lees_edwards/lees_edwards.hpp"
 
 #include "script_interface/ScriptInterface.hpp"
 #include "script_interface/auto_parameters/AutoParameters.hpp"
